+++ R/footnote.R
@@ -8,6 +8,13 @@
 #' and `j`. If you use i = c(1,3) and j = c(2,5), then you will
 #' add the symbols (or the repeated symbol) to cells `[1,2]`
 #' and `[3,5]`.
+#'
+#' \if{html}{\out{
+#' <img src="https://www.ardata.fr/img/flextable-imgs/flextable-016.png" alt="add_footer illustration" style="width:100\%;">
+#'
+#' See https://www.ardata.fr/en/flextable-gallery/2022-06-23-separate-headers/ for the example
+#' shown
+#' }}
 #' @param x a flextable object
 #' @param i,j cellwise rows and columns selection
 #' @param value a call to function [as_paragraph()].
@@ -19,33 +26,45 @@
 #' a separator between footnotes.
 #' @examples
 #' ft_1 <- flextable(head(iris))
-#' ft_1 <- footnote( ft_1, i = 1, j = 1:3,
-#'             value = as_paragraph(
-#'               c("This is footnote one",
-#'                 "This is footnote two",
-#'                 "This is footnote three")
-#'             ),
-#'             ref_symbols = c("a", "b", "c"),
-#'             part = "header")
+#' ft_1 <- footnote(ft_1,
+#'   i = 1, j = 1:3,
+#'   value = as_paragraph(
+#'     c(
+#'       "This is footnote one",
+#'       "This is footnote two",
+#'       "This is footnote three"
+#'     )
+#'   ),
+#'   ref_symbols = c("a", "b", "c"),
+#'   part = "header"
+#' )
 #' ft_1 <- valign(ft_1, valign = "bottom", part = "header")
 #' ft_1 <- autofit(ft_1)
 #'
 #' ft_2 <- flextable(head(iris))
 #' ft_2 <- autofit(ft_2)
-#' ft_2 <- footnote( ft_2, i = 1, j = 1:2,
-#'                value = as_paragraph(
-#'                 c("This is footnote one",
-#'                    "This is footnote two")
-#'                ),
-#'                ref_symbols = c("a", "b"),
-#'                part = "header", inline = TRUE)
-#' ft_2 <- footnote( ft_2, i = 1, j = 3:4,
-#'                value = as_paragraph(
-#'                  c("This is footnote three",
-#'                    "This is footnote four")
-#'                ),
-#'                ref_symbols = c("c","d"),
-#'                part = "header", inline = TRUE)
+#' ft_2 <- footnote(ft_2,
+#'   i = 1, j = 1:2,
+#'   value = as_paragraph(
+#'     c(
+#'       "This is footnote one",
+#'       "This is footnote two"
+#'     )
+#'   ),
+#'   ref_symbols = c("a", "b"),
+#'   part = "header", inline = TRUE
+#' )
+#' ft_2 <- footnote(ft_2,
+#'   i = 1, j = 3:4,
+#'   value = as_paragraph(
+#'     c(
+#'       "This is footnote three",
+#'       "This is footnote four"
+#'     )
+#'   ),
+#'   ref_symbols = c("c", "d"),
+#'   part = "header", inline = TRUE
+#' )
 #' ft_2
 #'
 #' ft_3 <- flextable(head(iris))
@@ -60,7 +79,7 @@
 #' @importFrom stats update
 footnote <- function(x, i = NULL, j = NULL, value, ref_symbols = NULL, part = "body",
                      inline = FALSE, sep = "; ") {
-  if( !inherits(x, "flextable") ) {
+  if (!inherits(x, "flextable")) {
     stop(sprintf("Function `%s` supports only flextable objects.", "footnote()"))
   }
 
@@ -82,10 +101,12 @@ footnote <- function(x, i = NULL, j = NULL, value, ref_symbols = NULL, part = "b
     symbols_str <- ref_symbols
   }
 
-  if(any(ref_symbols %in% "")) {
-    long_msg <- c("Usage of empty symbol '' with footnote should not happen, ",
-         "use `add_footer_lines()` instead, it does not require any symbol. ",
-         "This usage will be forbidden in the next release. Please, wait for 10 seconds!")
+  if (any(ref_symbols %in% "")) {
+    long_msg <- c(
+      "Usage of empty symbol '' with footnote should not happen, ",
+      "use `add_footer_lines()` instead, it does not require any symbol. ",
+      "This usage will be forbidden in the next release. Please, wait for 10 seconds!"
+    )
     long_msg <- paste0(long_msg, collapse = "\n")
     message(long_msg)
     Sys.sleep(10)
@@ -135,7 +156,6 @@ footnote <- function(x, i = NULL, j = NULL, value, ref_symbols = NULL, part = "b
     } else {
       x[["footer"]]$content[n_row, 1] <- list(rbind(x[["footer"]]$content[n_row, 1][[1]], paras))
     }
-
   } else {
     # init new lines
     x <- add_footer_lines(x, values = ref_symbols)
@@ -145,7 +165,7 @@ footnote <- function(x, i = NULL, j = NULL, value, ref_symbols = NULL, part = "b
       x <- prepend_chunks(
         x = x, i = n_row + v, j = 1,
         part = "footer",
-        as_sup(ref_symbols[v])# [ as we want a list of df
+        as_sup(ref_symbols[v]) # [ as we want a list of df
       )
     }
   }